implement the epoch logic

+++ x/bridge/types/query.pb.gw.go
@@ -1021,7 +1021,7 @@ func request_Query_GetEpochTransactions_0(ctx context.Context, marshaler runtime
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch_id")
 	}
 
-	protoReq.EpochId, err = runtime.Uint64(val)
+	protoReq.EpochId, err = runtime.Uint32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch_id", err)
@@ -1055,7 +1055,7 @@ func local_request_Query_GetEpochTransactions_0(ctx context.Context, marshaler r
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch_id")
 	}
 
-	protoReq.EpochId, err = runtime.Uint64(val)
+	protoReq.EpochId, err = runtime.Uint32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch_id", err)
@@ -1089,7 +1089,7 @@ func request_Query_GetEpochById_0(ctx context.Context, marshaler runtime.Marshal
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch_id")
 	}
 
-	protoReq.EpochId, err = runtime.Uint64(val)
+	protoReq.EpochId, err = runtime.Uint32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch_id", err)
@@ -1116,7 +1116,7 @@ func local_request_Query_GetEpochById_0(ctx context.Context, marshaler runtime.M
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "missing parameter %s", "epoch_id")
 	}
 
-	protoReq.EpochId, err = runtime.Uint64(val)
+	protoReq.EpochId, err = runtime.Uint32(val)
 
 	if err != nil {
 		return nil, metadata, status.Errorf(codes.InvalidArgument, "type mismatch, parameter: %s, error: %v", "epoch_id", err)
